Adding two roles: Artist and House
Fetlife got a space for a lot of artists to share their content with people. The Role "Artist" would make it easier to separate these profiles from the private ones. 

A lot of people are "Member of a House" with someone and a lot of places and organizations are using Fetlife to share information. The Term "House" would make it easier to separate human accounts from organizations.